tourism-backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
pymongo==4.5.0
nltk==3.8.1
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==2.3.0
